Minor fixes

+++ Proyectos/Acme Car'n Go/src/main/webapp/views/message/list.jsp
@@ -28,7 +28,7 @@
 
 	
 	<display:column>
-		<a href="message/actor/delete.do?messageId=${row.id}">
+		<a href="message/actor/delete.do?messageId=${row.id}" onclick="return confirm('<spring:message code="message.confirm.delete" />')" >
 			<spring:message code="message.delete" />
 		</a>
 	</display:column>
